requirements.txt
Flask==2.0.2
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.0
google-cloud-logging==2.7.0
google-cloud-error-reporting==1.4.1
gunicorn==20.1.0
six==1.16.0
